fix: simplify construction to speed up
