--- src/zeroconf/_logger.pxd
@@ -1,4 +1,0 @@
-
-cdef class QuietLogger:
-
-    cdef public object _seen_logs
