+++ src/zeroconf/_protocol/incoming.pxd
@@ -1,11 +1,10 @@
 
 import cython
 
-from .message cimport DNSMessage
-
 
 cdef class DNSIncoming:
 
+    cdef public object _seen_logs
     cdef public object offset
     cdef public object data
     cdef public object data_len
