--- src/zeroconf/_protocol/message.pxd
@@ -1,8 +1,0 @@
-
-
-from zeroconf._logger cimport QuietLogger
-
-
-cdef class DNSMessage:
-
-   cdef public object flags
